# contracts/care-coordination-platform-openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Patient Engagement Platform API
  version: 1.0.0
  description: API-first contract examples for the care coordination platform.
servers:
  - url: https://api.healthcare.example.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CarePlanRequest:
      type: object
      required: [patientId, goal, planStatus, ownerId, tasks]
      properties:
        patientId: { type: string }
        goal: { type: string }
        planStatus: { type: string }
        ownerId: { type: string }
        tasks:
          type: array
          items: { type: string }
      example:
        patientId: pat-1001
        goal: Improve glycemic control
        planStatus: ACTIVE
        ownerId: clinician-91
        tasks:
          - Track fasting glucose daily
          - Complete weekly nutrition review
    CarePlanUpdateRequest:
      type: object
      required: [goal, planStatus, ownerId, tasks, expectedVersion]
      properties:
        goal: { type: string }
        planStatus: { type: string }
        ownerId: { type: string }
        tasks:
          type: array
          items: { type: string }
        expectedVersion: { type: integer, minimum: 1 }
      example:
        goal: Improve glycemic control and mobility
        planStatus: ACTIVE
        ownerId: clinician-91
        tasks:
          - Track fasting glucose daily
          - Walk for 30 minutes
        expectedVersion: 1
    ConsentRequest:
      type: object
      required: [patientId, consentType, granted, effectiveFrom]
      properties:
        patientId: { type: string }
        consentType: { type: string }
        granted: { type: boolean }
        effectiveFrom: { type: string, format: date-time }
      example: { patientId: "pat-1001", consentType: "RESEARCH", granted: true, effectiveFrom: "2026-05-31T10:00:00Z" }
...
    DeviceIngestionRequest:
      type: object
      required: [deviceId, protocol, payload, receivedAt]
      properties:
        deviceId: { type: string }
        protocol:
          type: string
          enum: [MQTT, HTTP, AMQP]
        payload:
          type: string
          description: JSON string with at least one supported metric key (bp, hr, glucose, spo2, temperature)
        receivedAt: { type: string, format: date-time }
      example: { deviceId: "dev-22", protocol: "MQTT", payload: "{\"spo2\":97}", receivedAt: "2026-05-31T10:07:00Z" }
...
    CarePlanResponse:
      type: object
      required: [id, status, patientId, goal, planStatus, ownerId, tasks, version]
      properties:
        id: { type: string }
        status: { type: string }
        patientId: { type: string }
        goal: { type: string }
        planStatus: { type: string }
        ownerId: { type: string }
        tasks:
          type: array
          items: { type: string }
        version: { type: integer }
      example:
        id: cp-3001
        status: MANAGED
        patientId: pat-1001
        goal: Improve glycemic control
        planStatus: ACTIVE
        ownerId: clinician-91
        tasks:
          - Track fasting glucose daily
          - Complete weekly nutrition review
        version: 1
    CarePlanResponsibilityResponse:
      type: object
      required: [patientId, ownerId, carePlanId, planStatus, version]
      properties:
        patientId: { type: string }
        ownerId: { type: string }
        carePlanId: { type: string }
        planStatus: { type: string }
        version: { type: integer }
      example:
        patientId: pat-1001
        ownerId: clinician-91
        carePlanId: cp-3001
        planStatus: ACTIVE
        version: 2
    ConsentResponse:
      type: object
      required: [id, status, patientId, consentType, granted, effectiveFrom]
      properties:
        id: { type: string }
        status: { type: string }
        patientId: { type: string }
        consentType: { type: string }
        granted: { type: string }
        effectiveFrom: { type: string }
      example: { id: "con-401", status: "RECORDED", patientId: "pat-1001", consentType: "RESEARCH", granted: "true", effectiveFrom: "2026-05-31T10:00:00Z" }
...
    DeviceIngestionResponse:
      type: object
      required: [id, status, deviceId, protocol, payload, receivedAt]
      properties:
        id: { type: string }
        status: { type: string }
        deviceId: { type: string }
        protocol: { type: string }
        payload: { type: string }
        receivedAt: { type: string }
      example: { id: "ing-1001", status: "INGESTED", deviceId: "dev-22", protocol: "MQTT", payload: "{\"spo2\":97}", receivedAt: "2026-05-31T10:07:00Z" }
...
    CarePlanEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/CarePlanResponse' }
    CarePlanListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/CarePlanResponse' }
    CarePlanResponsibilityEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/CarePlanResponsibilityResponse' }
    ConsentEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/ConsentResponse' }
    ConsentListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/ConsentResponse' }
...
    DeviceIngestionEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data: { $ref: '#/components/schemas/DeviceIngestionResponse' }
    DeviceIngestionListEnvelope:
      type: object
      required: [correlationId, data]
      properties:
        correlationId: { type: string }
        data:
          type: array
          items: { $ref: '#/components/schemas/DeviceIngestionResponse' }
...
    ApiError:
      type: object
      required: [correlationId, code, message, timestamp]
      properties:
        correlationId: { type: string }
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }
      example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Patient record not found: pat-404", timestamp: "2026-05-31T10:10:00Z" }
  responses:
    BadRequestResponse:
      description: Validation or request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "VALIDATION_ERROR", message: "Validation failed", timestamp: "2026-05-31T10:11:00Z" }
    NotFoundResponse:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "RESOURCE_NOT_FOUND", message: "Resource not found", timestamp: "2026-05-31T10:11:20Z" }
    ConflictResponse:
      description: Business rule conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "SLOT_ALREADY_BOOKED", message: "Slot is already booked", timestamp: "2026-05-31T10:11:30Z" }
    InternalServerErrorResponse:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: { correlationId: "corr-123", code: "INTERNAL_ERROR", message: "Unexpected server error", timestamp: "2026-05-31T10:11:40Z" }
security:
  - bearerAuth: []
paths:
...
  /careplans:
    post:
      summary: Create a care plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarePlanRequest'
            examples:
              createCarePlan:
                value:
                  patientId: pat-1001
                  goal: Improve glycemic control
                  planStatus: ACTIVE
                  ownerId: clinician-91
                  tasks:
                    - Track fasting glucose daily
                    - Complete weekly nutrition review
      responses:
        '201':
          description: Care plan created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarePlanEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: cp-3001
                      status: MANAGED
                      patientId: pat-1001
                      goal: Improve glycemic control
                      planStatus: ACTIVE
                      ownerId: clinician-91
                      tasks:
                        - Track fasting glucose daily
                        - Complete weekly nutrition review
                      version: 1
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List care plans
      responses:
        '200':
          description: Care plan collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarePlanListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: cp-3001
                        status: MANAGED
                        patientId: pat-1001
                        goal: Improve glycemic control
                        planStatus: ACTIVE
                        ownerId: clinician-91
                        tasks:
                          - Track fasting glucose daily
                          - Complete weekly nutrition review
                        version: 1
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /careplans/{id}:
    put:
      summary: Update a care plan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarePlanUpdateRequest'
            examples:
              updateCarePlan:
                value:
                  goal: Improve glycemic control and mobility
                  planStatus: ACTIVE
                  ownerId: clinician-91
                  tasks:
                    - Track fasting glucose daily
                    - Walk for 30 minutes
                  expectedVersion: 1
      responses:
        '200':
          description: Care plan updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarePlanEnvelope' }
              examples:
                updated:
                  value:
                    correlationId: corr-123
                    data:
                      id: cp-3001
                      status: MANAGED
                      patientId: pat-1001
                      goal: Improve glycemic control and mobility
                      planStatus: ACTIVE
                      ownerId: clinician-91
                      tasks:
                        - Track fasting glucose daily
                        - Walk for 30 minutes
                      version: 2
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '409': { $ref: '#/components/responses/ConflictResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: Retrieve a care plan
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Care plan retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarePlanEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: cp-3001
                      status: MANAGED
                      patientId: pat-1001
                      goal: Improve glycemic control
                      planStatus: ACTIVE
                      ownerId: clinician-91
                      tasks:
                        - Track fasting glucose daily
                        - Complete weekly nutrition review
                      version: 1
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /careplans/responsibility/{patientId}:
    get:
      summary: Retrieve care coordinator responsibility for a patient
      parameters:
        - in: path
          name: patientId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Responsibility retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CarePlanResponsibilityEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      patientId: pat-1001
                      ownerId: clinician-91
                      carePlanId: cp-3001
                      planStatus: ACTIVE
                      version: 2
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /consents:
    post:
      summary: Record consent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
            examples:
              recordConsent:
                value:
                  patientId: pat-1001
                  consentType: RESEARCH
                  granted: true
                  effectiveFrom: 2026-05-31T10:00:00Z
      responses:
        '201':
          description: Consent recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsentEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: con-401
                      status: RECORDED
                      patientId: pat-1001
                      consentType: RESEARCH
                      granted: "true"
                      effectiveFrom: 2026-05-31T10:00:00Z
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List consents
      responses:
        '200':
          description: Consent collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsentListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: con-401
                        status: RECORDED
                        patientId: pat-1001
                        consentType: RESEARCH
                        granted: "true"
                        effectiveFrom: 2026-05-31T10:00:00Z
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /consents/{id}:
    get:
      summary: Retrieve a consent
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Consent retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsentEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: con-401
                      status: RECORDED
                      patientId: pat-1001
                      consentType: RESEARCH
                      granted: "true"
                      effectiveFrom: 2026-05-31T10:00:00Z
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...
  /devices/events:
    post:
      summary: Ingest device signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceIngestionRequest'
            examples:
              ingestDeviceSignal:
                value:
                  deviceId: dev-22
                  protocol: MQTT
                  payload: '{"spo2":97}'
                  receivedAt: 2026-05-31T10:07:00Z
      responses:
        '201':
          description: Device signal ingested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceIngestionEnvelope' }
              examples:
                created:
                  value:
                    correlationId: corr-123
                    data:
                      id: ing-1001
                      status: INGESTED
                      deviceId: dev-22
                      protocol: MQTT
                      payload: '{"spo2":97}'
                      receivedAt: 2026-05-31T10:07:00Z
        '400': { $ref: '#/components/responses/BadRequestResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
    get:
      summary: List ingested device events
      responses:
        '200':
          description: Device event collection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceIngestionListEnvelope' }
              examples:
                list:
                  value:
                    correlationId: corr-123
                    data:
                      - id: ing-1001
                        status: INGESTED
                        deviceId: dev-22
                        protocol: MQTT
                        payload: '{"spo2":97}'
                        receivedAt: 2026-05-31T10:07:00Z
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
  /devices/events/{id}:
    get:
      summary: Retrieve ingested device event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Device event retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceIngestionEnvelope' }
              examples:
                found:
                  value:
                    correlationId: corr-123
                    data:
                      id: ing-1001
                      status: INGESTED
                      deviceId: dev-22
                      protocol: MQTT
                      payload: '{"spo2":97}'
                      receivedAt: 2026-05-31T10:07:00Z
        '404': { $ref: '#/components/responses/NotFoundResponse' }
        '500': { $ref: '#/components/responses/InternalServerErrorResponse' }
...